Refactor train function to Trainable base class
Changed HyperParamsClass to be Trainable class. Moved the train function
inside Trainable. Realized this is better because the class already had
all the necessary train hyperparameters, that we might as well just make
it train itself.

This will also make it easier to add more model types by extending Trainable